+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(unused_crate_dependencies, warnings)]
 
+mod duckdb;
 mod error;
 mod migrate;
 mod read;
@@ -7,7 +8,7 @@ mod search;
 mod version;
 mod write;
 
-use duckdb as _;
+use ::duckdb as _;
 use error::Error;
 use pyo3::prelude::*;
 
@@ -16,6 +17,7 @@ type Result<T> = std::result::Result<T, Error>;
 /// A collection of functions for working with STAC, using Rust under the hood.
 #[pymodule]
 fn stacrs(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<duckdb::DuckdbClient>()?;
     m.add_function(wrap_pyfunction!(migrate::migrate, m)?)?;
     m.add_function(wrap_pyfunction!(migrate::migrate_href, m)?)?;
     m.add_function(wrap_pyfunction!(read::read, m)?)?;